requirements.txt
fastapi
uvicorn
requests
fpdf2
python-multipart
python-dotenv